Consolidate link click webhook sync and centralize WebhookTrigger type
