+++ apps/web/lib/webhook/publish.ts
@@ -1,8 +1,8 @@
 import { prisma } from "@/lib/prisma";
 import { Webhook } from "@prisma/client";
-import { WebhookTrigger, WorkspaceProps } from "../types";
+import { WorkspaceProps } from "../types";
 import { sendWebhooks } from "./qstash";
-import { WebhookEventPayload } from "./types";
+import type { WebhookEventPayload, WebhookTrigger } from "./types";
 
 // Send workspace level webhook
 export const sendWorkspaceWebhook = async ({
