+++ apps/web/lib/webhook/qstash.ts
@@ -6,7 +6,7 @@ import * as z from "zod/v4";
 import { formatEventForSegment } from "../integrations/segment/transform";
 import { createSegmentBasicAuthHeader } from "../integrations/segment/utils";
 import { formatEventForSlack } from "../integrations/slack/transform";
-import { WebhookTrigger } from "../types";
+import type { WebhookTrigger } from "./types";
 import { createWebhookSignature } from "./signature";
 import { prepareWebhookPayload } from "./transform";
 import { WebhookEventPayload } from "./types";
