+++ apps/web/lib/webhook/types.ts
@@ -5,11 +5,14 @@ import { linkEventSchema } from "../zod/schemas/links";
 import { EnrolledPartnerSchema } from "../zod/schemas/partners";
 import { payoutWebhookEventSchema } from "../zod/schemas/payouts";
 import { partnerApplicationWebhookSchema } from "../zod/schemas/program-application";
+import { WEBHOOK_TRIGGERS } from "./constants";
 import {
   clickWebhookEventSchema,
   leadWebhookEventSchema,
   saleWebhookEventSchema,
 } from "./schemas";
+
+export type WebhookTrigger = (typeof WEBHOOK_TRIGGERS)[number];
 
 export type ClickEventWebhookPayload = z.infer<typeof clickWebhookEventSchema>;
