+++ apps/web/ui/modals/send-test-webhook-modal.tsx
@@ -1,6 +1,7 @@
 import { sendTestWebhookEvent } from "@/lib/actions/send-test-webhook";
 import useWorkspace from "@/lib/swr/use-workspace";
-import { WebhookProps, WebhookTrigger } from "@/lib/types";
+import { WebhookProps } from "@/lib/types";
+import type { WebhookTrigger } from "@/lib/webhook/types";
 import { WEBHOOK_TRIGGER_DESCRIPTIONS } from "@/lib/webhook/constants";
 import { Button, Combobox, ComboboxOption, Modal } from "@dub/ui";
 import { useAction } from "next-safe-action/hooks";
